// src/features/course-v1/model/types.ts
export interface CoursePoint {
  pointSeq: number;
  lat: number;
  lng: number;
  ele: number;
}